refactored all utility funcs into utils file
